qimgv/gui/dialogs/文件夹下全部文件优化 printdialog - updatePreview(printdialog.cpp:111-118):缩放结果按目标像素尺寸缓存(新增成员 cachedScaledImg,printdialog.h:41),几何不变时复用,不再每次切控件都对大图平滑重采样;setImage 时失效缓存。 - setupConnections(printdialog.cpp:74):连接从 color 改到 grayscale——修复切灰度预览不刷新的 bug,同时去掉 color 切换导致的无效重缩放。 - exportPdf(printdialog.cpp:179-184):删掉恒为假的 outputFileName().isEmpty() 判空,只按 !img 判断。 settingsdialog - saveSettings(settingsdialog.cpp:273-275):移除 waitForDone() 阻塞全局线程池。后台 ScalerRunnable 只读一次 smoothUpscaling,等待会让关设置卡在无关的图片任务上。 - 语言回退(settingsdialog.cpp:246):setCurrentText("en_US") → langs.value("en_US")("English"),修好 "en_US" 不在列表导致回退失效的问题。 - readSettings L227-239 的显式 handler 调用保留了——它们是防御性的(值未变化时 setValue 不发信号),删掉有初始 label 不同步的风险。 scripteditordialog - initializeEditMode:onNameChanged 的第一次空值调用已删(scripteditordialog.cpp:35-43),保留 setText 后那次保证空名边界。 - selectScriptPath(:83-88):去掉未使用的栈上 QFileDialog dialog,改用静态 QFileDialog::getOpenFileName。 resizedialog:widthChanged/heightChanged(:79-93)的 float 统一为 double。 shortcutcreatordialog:setAction 改 const QString&(头文件同步更新),仅必要时拷贝。 filereplacedialog:bool multi 改为头文件初始化 = false,printdialog.cpp:6 的赋值删除。

+++ qimgv/gui/dialogs/resizedialog.cpp
@@ -78,17 +78,17 @@ QSize ResizeDialog::newSize() {
 
 void ResizeDialog::widthChanged(int newWidth) {
     // 始终按原始比例计算高度
-    float factor = static_cast<float>(newWidth) / static_cast<float>(originalSize.width());
+    double factor = static_cast<double>(newWidth) / static_cast<double>(originalSize.width());
     targetSize.setWidth(newWidth);
-    targetSize.setHeight(static_cast<int>(static_cast<float>(originalSize.height()) * factor));
+    targetSize.setHeight(static_cast<int>(static_cast<double>(originalSize.height()) * factor));
     updateToTargetValues();
 }
 
 void ResizeDialog::heightChanged(int newHeight) {
     // 始终按原始比例计算宽度
-    float factor = static_cast<float>(newHeight) / static_cast<float>(originalSize.height());
+    double factor = static_cast<double>(newHeight) / static_cast<double>(originalSize.height());
     targetSize.setHeight(newHeight);
-    targetSize.setWidth(static_cast<int>(static_cast<float>(originalSize.width()) * factor));
+    targetSize.setWidth(static_cast<int>(static_cast<double>(originalSize.width()) * factor));
     updateToTargetValues();
 }
